Equalizer Preset - Add and Remove now separate tests
